feat: ensure only authenticated user can book
